replace logic to write data to SettPhotos table
so that uploaded file data correctly associated with each sett.

Also extracted out email sending code into separate email module,
and de-coupled it from database writing code so that the returns
data is still written to the DB even if the email sending fails.

Issue: Scottish-Natural-Heritage/Licensing#1497
